auto_repair_shop/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.0.12
Django==3.2.16
django-bootstrap5==23.1
factory-boy==3.2.1
Faker==18.9.0
idna==3.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
